ai_service/requirements.txt
# =============================================================
# 基础依赖（必装）：即便不装 AI 模型，服务也能以 mock 模式启动
# =============================================================
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
numpy==1.26.4
Pillow==10.4.0

# 通义千问（推荐装，API 调用文本分析）
dashscope==1.20.3

# =============================================================
# 真实 AI 模型依赖（可选）：如需使用真实推理，取消对应注释
# =============================================================
# ----- CV 基础 -----
# opencv-python-headless==4.10.0.84
# torch==2.3.1
# torchvision==0.18.1

# ----- 行为检测：YOLOv8 -----
# ultralytics==8.2.79

# ----- 人脸识别：InsightFace -----
# insightface==0.7.3
# onnxruntime-gpu==1.18.1   # 开发机有 GPU 用这个
# onnxruntime==1.18.1       # 纯 CPU 机器用这个

# ----- 表情识别：HSEmotion -----
# hsemotion==0.3
# timm==1.0.8

# ----- 姿态识别（M3 阶段会用到）-----
# mediapipe==0.10.14

# ----- 语音转文字（M4 阶段可选）-----
# openai-whisper==20231117
